카카오맵정보DB처리/구글차트연습

+++ src/main/java/com/spring/cjs2108/service/AdminServiceImpl.java
@@ -49,10 +49,13 @@ public class AdminServiceImpl implements AdminService {
 		File path = new File(uploadPath);
 		// 파일객체를 통해서 uploadPath경로안의 모든 파일의 정보를 담아와서 배열로 저장한다.
 		File[] fileList = path.listFiles();
-		int fileCnt = fileList.length - 1;
+		int fileCnt = 0;
 		
-		for(int i=0; i<fileCnt; i++) {
-			fileList[i].delete();
+		for(int i=0; i<fileList.length; i++) {
+			if(fileList[i].exists() && !fileList[i].isDirectory()) {
+				fileList[i].delete();
+				fileCnt++;
+		  }
 		}
 		return fileCnt;
 	}
